destorys the user_beers resource and sets up the scripts for the beers
got ride of all of the migrations, controllers, modela, routes ect for
the user_beers table then added a migration on to have the beers table
have a one to many relationship with users. then nuke and paved the db
so that the only two resources now are beers and users. Also sets up the
curl scripts to curl the db with no protection on the db yet. either from
the current user or a login token